+++ README.md
@@ -1,2 +1,39 @@
 # Astar
-AI Principles
+
+This repository contains implementations of various pathfinding algorithms such as Uniform Cost Search (UCS), A* Search, and Dijkstra's Algorithm. The code is organized to facilitate collaboration and reuse of common functionalities.
+
+## Table of Contents
+
+- [Astar](#astar)
+  - [Table of Contents](#table-of-contents)
+  - [Installation](#installation)
+  - [Usage](#usage)
+
+## Installation
+
+1. Clone the repository:
+
+    ```sh
+    git clone https://github.com/yourusername/pathfinding-algorithms.git
+    cd pathfinding-algorithms
+    ```
+
+2. Create a virtual environment:
+
+    ```sh
+    python3 -m venv venv
+    source venv/bin/activate  # On Windows use `venv\Scripts\activate`
+    ```
+
+3. Install the required dependencies:
+
+    ```sh
+    pip install -r requirements.txt
+    ```
+
+## Usage
+
+To run a specific algorithm, use the `main.py` script. For example, to run the Uniform Cost Search (UCS) algorithm:
+
+```sh
+python [main.py](http://_vscodecontentref_/1)
