+++ main.py
@@ -1,0 +1,7 @@
+from algorithms import ucs
+
+def main():
+    ucs.main()
+
+if __name__ == "__main__":
+    main()
